.info is broken.
Underlying package requires no compilation, just installing a script
and an etc/ config file (for which the location must be patched into
the script) (and a ConfFiles declared) and copying the manpages and
other docs.

This pacakge may never be ready for Fink, as it relies on the names of
softlinks pointing to the actual script to figure out what to do, and
those names are common system commands. I'm not gonna add a bunch of
non-standard-behavior actions for standard commands, so will probably
put them all in their own lib directory and then add it to $PATH (via
profile.d scripts).