New upstream release;
0.5.1
The changes in this release are as follows:
This release removes memory leaks in generated C files.

0.5
The changes in this release are as follows:
C code can be generated instead of C++ code. If an item is of type
"method" it can still be assigned a string and in that case the
method is not invoked (the same holds for C external functions). The
indent counter is correctly reset to 0 when it encounters a new line
and no spaces at the beginning of the line. An --expand-tabs option
has been added that allows you to correctly indent strings with
newlines inside.