New upstream release;
3.0.1
The changes in this release are as follows:
Timeout handling in SFTP has been fixed. A core dump in the case of
some network errors has been fixed. There are several compilation
fixes.

3.0.0
The changes in this release are as follows:
sftp protocol support was added along with support for  advanced FTP
features such as FEAT, LANG, and MLSD.  There were also mirror
improvements and other minor  feature additions and bugfixes.