New upstream release;
3.0.2
The changes in this release are as follows:
Support has been added for the CLNT, HOST, and OPTS MLST FTP commands.
MLSD is disabled by default. Several bugs have been fixed.

3.0.3
The changes in this release are as follows:
cls for was sftp improved. 'cd dir/' now checks if the directory
exists, and ftp:ssl-force is honored even if the FTP server  does not
advertise AUTH in its FEAT reply. Several bugs were  fixed in FXP and
SFTP.

3.0.4
The changes in this release are as follows:
SSL-encrypted FXP support was added. CPU usage optimizations were
made. Some workarounds for broken FTP servers were added. Some bugs
were fixed.