Even more perl packages, en route to a full ikiwiki package.
I don't care about any of these packages, per se; I just want to get a
fully-featured ikiwiki fink package, which means adding all its
dependencies, of which there are a fair number.  I also want to ensure
that the optional components of ikiwiki work.  This process has
revealed some holes in fink's coverage of CPAN, and I aim to fix that.

I'm committing this batch, which has been under construction for a
week or so, just to get my work out in the open.  If you're working on
any of these packages, please let me know so we don't step on each
other's toes.  If you have access to a machine running 10.5 or 10.4
and are bored, I'd appreciate finding out whether these packages build
successfully.