More perl goodness.
The menagerie of dependencies for ikiwiki keeps getting more and more
menagerie-y.  There are still several more to go before a totally
complete ikiwiki package is ready.

Please tell me if you're working on any of these perl modules.  If
you're bored and have a 10.5 or 10.4 machine, give these a whirl.